moved main logic to useDataSaver
